fix: fix segmentation fault fix #2

+++ tool/lemon.c
@@ -1322,7 +1322,7 @@ char **argv;
 /*
 ** Return a pointer to the next structure in the linked list.
 */
-#define NEXT(A) (*(char**)(((int)A)+offset))
+#define NEXT(A) (*(char**)(((size_t)A)+offset))
 
 /*
 ** Inputs:
